Add new coordinator for profiler questions
